# pkg/docs/swagger.yaml
definitions:
  models.Communication:
    properties:
      login:
        type: string
      messenger:
        items:
          $ref: '#/definitions/models.Messenger'
        type: array
      parentId:
        description: gorm.Model
        type: integer
    type: object
  models.Education:
    properties:
      degree:
        type: string
      endYear:
        type: integer
      institution:
        type: string
      parentId:
        description: gorm.Model
        type: integer
      startYear:
        type: integer
    type: object
  models.Messenger:
    properties:
      communications:
        items:
          $ref: '#/definitions/models.Communication'
        type: array
      name:
        description: gorm.Model
        type: string
    type: object
  models.OtherInformation:
    properties:
      data:
        type: string
      parentId:
        description: gorm.Model
        type: integer
    type: object
  models.Report:
    properties:
      fromUserId:
        description: gorm.Model
        type: integer
      reportText:
        type: string
      reportTheme:
        type: string
      toUserId:
        type: integer
    type: object
  models.User:
    properties:
      averageClassPrice:
        type: integer
      communications:
        items:
          $ref: '#/definitions/models.Communication'
        type: array
      dateOfBirthday:
        type: string
      description:
        type: string
      educations:
        items:
          $ref: '#/definitions/models.Education'
        type: array
      email:
        description: |-
          gorm.Model
          Base information
        type: string
      firstName:
        type: string
      isMentor:
        type: boolean
      isVerifyEmail:
        type: boolean
      isVerifyPhone:
        type: boolean
      otherInformation:
        items:
          $ref: '#/definitions/models.OtherInformation'
        type: array
      patronymic:
        type: string
      phone:
        type: string
      profilePicture:
        type: string
      rating:
        type: number
      reports:
        description: Жалобы на пользователя
        items:
          $ref: '#/definitions/models.Report'
        type: array
      secondName:
        type: string
      specialization:
        description: Mentor information
        type: string
      time:
        type: string
      userReports:
        description: Жалобы пользователя
        items:
          $ref: '#/definitions/models.Report'
        type: array
      workExperiences:
        items:
          $ref: '#/definitions/models.WorkExperience'
        type: array
    type: object
  models.WorkExperience:
    properties:
      endYear:
        type: string
      organization:
        type: string
      parentId:
        description: gorm.Model
        type: integer
      startYear:
        type: string
    type: object
info:
  contact: {}
  description: Clients methods for skipper cms
  title: Clients service
  version: "1.0"
paths:
  /clients/:
    get:
      consumes:
      - application/json
      description: Получение списка менторов и менти
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Clients
  /clients/{ClientId}:
    get:
      consumes:
      - application/json
      description: Получение конкретного пользователя
      parameters:
      - description: Client ID
        in: path
        name: ClientId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerAuth: []
      tags:
      - Clients
swagger: "2.0"